feat: replace all monetary f64 with Decimal across codebase
- models.rs: Transaction.amount and ParsedRow.amount now Decimal
- db.rs: amount/balance columns changed from REAL to TEXT
- importer.rs: parse_amount returns Decimal, all SQL params use .to_string()
- reports.rs: all report structs use Decimal, added read_decimal/read_decimal_sum helpers
- reconciler.rs: ReconcileResult fields and reconcile() param use Decimal
- reviewer.rs: FlaggedTxn.amount uses Decimal
- browser.rs: RegisterBrowser.total uses Decimal
- dashboard.rs: HomeData monetary fields use Decimal, ToPrimitive for chart values
- cli/mod.rs + reconcile.rs: CLI balance arg is Decimal
- reconcile_manager.rs: local result fields use Decimal
- report/view.rs + text.rs: Decimal comparisons throughout
- pdf.rs: Decimal comparisons for color logic
- demo.rs: all demo amounts use Decimal (from_parts for consts, Decimal::new for runtime)
- snake.rs: food_value and score use Decimal
- All 157 tests updated and passing

+++ src/cli/mod.rs
@@ -25,6 +25,7 @@ pub mod splash;
 pub mod status;
 
 use clap::{Args, Parser, Subcommand};
+use rust_decimal::Decimal;
 
 pub(crate) fn parse_month_opt(month: &Option<String>) -> (Option<i32>, Option<u32>) {
     if let Some(m) = month {
@@ -103,7 +104,7 @@ pub enum Commands {
         month: String,
         /// Statement ending balance
         #[arg(long)]
-        balance: f64,
+        balance: Decimal,
     },
     /// Switch to an existing Nigel data directory.
     Load {
